requirements.txt
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.0
cycler==0.12.1
dash==2.16.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
Data-Generator==1.0.1
exceptiongroup==1.2.0
Flask==3.0.2
fonttools==4.49.0
idna==3.6
importlib-metadata==7.0.1
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.3
kafka-python==2.0.2
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.8.3
nest-asyncio==1.6.0
numpy==1.26.4
packaging==23.2
pandas==2.2.1
pillow==10.2.0
plotly==5.19.0
pluggy==1.4.0
pyparsing==3.1.1
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
pytest==8.0.2
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.31.0
retrying==1.3.4
six==1.16.0
tenacity==8.2.3
tomli==2.0.1
tomlkit==0.12.4
tqdm==4.66.2
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
Werkzeug==3.0.1
XlsxWriter==3.2.0
zipp==3.17.0
